Version 0.6.3 11-06-2025 16:55
Updated theme logic to read/write from/to Firebase.
Updated log-in and log-out logic to properly reflect preferences.
Changed Card logic to fade-in on log-in and stay visible until user logs-out or the app is closed.
Added a bunch of debug lines.

Have to create a debug switch.
Other preferences/settings need to be moved to Firebase.